Revert "Removes unused (/useless?) 'baseinterpreter' type & params"
This reverts commit 73b936f4a0a6e3b38e41f9cbf8bbeaeecfd37a9b
which removed the 'baseinterpreter' type parameter to many interpreter
functions. It turns out it was usefull after all.
For example, if an instruction refers to the type of the entire computation as
in:

data ConcurrentInst p a where
  Fork :: p () -> ConcurrentInst p ()

Then without threading the baseinterpreter p through the interpreter composition
functions, p can only be evaluated if it's constrained in some other way. When
removed, this wasnt noticed as code used to test it statically knew that p was
equal to 'Program (Inst1 :+: Inst2)' and could make use of this knowledge.

By passing this parameter, 'CoreInst' can be contained in some unknown sum and
still evaluate p.